add function to drop the previous screenshots from historical messages
